Began Nearest Neighbor Search Algerithm V2
Started working on the second iteration of the nearest neighbor search algorithm, this time using a cache of neighboring cells to traverse the tree. So far, it's mostly working, although there are still a few bugs, so it's not quite ready. The old nearest neighbor search is now labeled ApproximateNearestNeighborSearch.